requirements.txt
# core
langchain>=0.1.0
openai>=0.27.0
playwright>=1.48.0
beautifulsoup4>=4.12.3
requests>=2.28.0

# vector DB
chromadb>=0.4.0
# langchain chroma adapter (if older langchain, use langchain-vectorstores)
langchain-chroma>=0.1.0

# embeddings
# using OpenAI embeddings here; substitute if you want other providers
# (if using local embeddings, add their libs)

# utilities
python-dotenv>=1.0.0
tqdm

# --- LangChain Ecosystem ---
langchain==0.3.8
langchain-core==0.3.35
langchain-community==0.3.7
langchain-openai==0.3.5
langchain-text-splitters==0.3.4

# --- Ecosystem ---
langsmith==0.1.140
openai>=1.58.1

# --- Compatibility Dependencies ---
pydantic==2.9.2
pydantic-settings==2.11.0
tenacity==8.5.0
tiktoken==0.8.0
typing-extensions>=4.9.0
